+++ mw/preference.ts
@@ -149,7 +149,7 @@ export const openPreference = async (editors: (CodeMirror | undefined)[]): Promi
 		themeWidget.setValue(theme);
 	} else {
 		dialog = new OO.ui.MessageDialog({id: 'cm-preference'});
-		dialog.$element.css('z-index', '801');
+		dialog.$element.css('z-index', 801);
 		const windowManager = new OO.ui.WindowManager();
 		windowManager.$element.appendTo(document.body);
 		windowManager.addWindows([dialog]);